Move code to Functions/ggmlEvaluate; Some ugly division into files and classes; Add storage of loaded models to Context
