+++ src/Interpreters/Context.cpp
@@ -67,6 +67,7 @@
 #include <Dictionaries/Embedded/GeoDictionariesLoader.h>
 #include <Interpreters/EmbeddedDictionaries.h>
 #include <Interpreters/ExternalDictionariesLoader.h>
+#include <Functions/ggmlEvaluate/gpt_common.h>
 #include <Functions/UserDefined/ExternalUserDefinedExecutableFunctionsLoader.h>
 #include <Functions/UserDefined/IUserDefinedSQLObjectsStorage.h>
 #include <Functions/UserDefined/createUserDefinedSQLObjectsStorage.h>
@@ -422,6 +423,8 @@ struct ContextSharedPart : boost::noncopyable
     mutable std::mutex keeper_dispatcher_mutex;
     mutable std::shared_ptr<KeeperDispatcher> keeper_dispatcher TSA_GUARDED_BY(keeper_dispatcher_mutex);
 #endif
+
+    GptStorage gpt_storage;
 
     ContextSharedPart()
         : access_control(std::make_unique<AccessControl>())
@@ -5391,6 +5394,11 @@ void Context::setClientProtocolVersion(UInt64 version)
 const ServerSettings & Context::getServerSettings() const
 {
     return shared->server_settings;
+}
+
+GptStorage & Context::getGptStorage() const
+{
+    return shared->gpt_storage;
 }
 
 }
